Move `schedule()` from `scheduler` into `task`
* Allows the `task` crate itself to directly invoke `schedule()`,
  which is important for upcoming efficiency improvements to
  regular and async task management (e.g., waking a task that was
  blocked on another task for the purpose of `join`).

* Allows future PRs to avoid cyclic dependencies when a crate
  depends on both task and scheduling functionality.

* The `scheduler` crate still provides the same scheduling functionality
  by re-exporting the same function from `task`.
