+++ kernel/task/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 log = "0.4.8"
 spin = "0.9.4"
 static_assertions = "1.1.0"
+crossbeam-utils = { version = "0.8.2", default-features = false }
 
 irq_safety = { git = "https://github.com/theseus-os/irq_safety" }
